chore: rename GameStateContext to jsx
