+++ client/Game.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { GameStateContext } from './GameStateContext.js';
+import { GameStateContext } from './GameStateContext.jsx';
 import { PHASES } from '../shared/constants.js';
 import Board from './Board.jsx';
 import PlayerList from './PlayerList.jsx';
