+++ client/PolicyHand.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { GameStateContext } from './GameStateContext.js';
+import { GameStateContext } from './GameStateContext.jsx';
 import { MESSAGE_TYPES } from '../shared/messages.js';
 
 /**
